Move resolution of dataset aliases (e.g. /flu/seasonal) into Dataset instances
One small step closer to a better separation of concerns and ultimately
removal of global.availableDatasets.  More immediately, this enables
resolution of aliases in a routing context that doesn't rely on the
"prefix"-handling functions.

parsePrefix()'s API doesn't change; it still returns a fully-resolved
Dataset instance and a resolvedPrefix for comparison.  Other usages to
follow will directly compare Dataset instances, however.
